requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pyjwt>=2.8.0
cryptography>=41.0.0
requests>=2.31.0
stripe>=7.0.0

# AI/ML (for agent capabilities)
openai>=1.0.0
anthropic>=0.8.0

# Optional: for advanced features
websockets>=12.0
aiofiles>=23.0.0
structlog>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
